[image_int] Move internal functions into separate translation unit
